requirements.txt
scrapy~=2.6
scrapy_user_agents~=0.1
requests~=2.27.1
pymysql~=1.0.2
datetime~=4.5
psycopg2-binary~=2.9.3
python-dateutil~=2.8.2